requirements.txt
streamlit
cohere
pinecone-client
PyPDF2
